+++ docs/changelog.md
@@ -15,7 +15,9 @@ pages](https://github.com/dalibo/ldap2pg/pulls?utf8=%E2%9C%93&q=is%3Apr%20is%3Am
 - Fails if configuration file is not found.
 - Refuse empty configuration file.
 - Refuse undefined `sync_map`.
-- Update sample configuration for readability and general use.
+- Update [sample
+  ldap2pg.yml](https://github.com/dalibo/ldap2pg/bloc/master/ldap2pg.yml) for
+  readability and general use.
 
 
 # ldap2pg 5.2