+++ docs/changelog.md
@@ -12,6 +12,7 @@ pages](https://github.com/dalibo/ldap2pg/pulls?utf8=%E2%9C%93&q=is%3Apr%20is%3Am
 - Fix grant to capitalized role.
 - Fix rename of members.
 - Log role after their originated LDAP query.
+- Add `description:` to mapping for logging.
 
 
 # ldap2pg 5.3