+++ docs/changelog.md
@@ -7,6 +7,14 @@ follow [merged Pull request
 pages](https://github.com/dalibo/ldap2pg/pulls?utf8=%E2%9C%93&q=is%3Apr%20is%3Amerged).
 
 
+# Unreleased
+
+- Fix grant to capitalized role.
+- Fix rename of members.
+- Log role after their originated LDAP query.
+- Add `description:` to mapping for logging.
+
+
 # ldap2pg 5.3
 
 - Fix join when multiple entries are returned.